+++ R/serialize.R
@@ -26,3 +26,33 @@ zstd_unserialize <- function(raw_vec) {
   .Call('zstd_unserialize_', raw_vec)
 }
 
+
+
+
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+#' Serialize arbitrary objects to a compressed stream of bytes using Zstandard
+#'
+#' @param raw_vec raw vector
+#' @param level compression level -5 to 22. Default: 3
+#'
+#' @return raw vector
+#'
+#' @export
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+zstd_compress <- function(raw_vec, level = 3L) {
+  .Call('zstd_compress_', raw_vec, level)
+}
+
+
+
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+#' Decompress raw bytes
+#' 
+#' @param raw_vec raw vector 
+#' 
+#' @export
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+zstd_decompress <- function(raw_vec) {
+  .Call('zstd_decompress_', raw_vec)
+}
+